+++ stovepipe/extension/buildrunner/buildkite/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
     importpath = "github.com/uber/submitqueue/stovepipe/extension/buildrunner/buildkite",
     visibility = ["//visibility:public"],
     deps = [
+        "//platform/buildkite:go_default_library",
         "//platform/errs:go_default_library",
-        "//platform/extension/buildrunner/buildkite:go_default_library",
         "//platform/http:go_default_library",
         "//stovepipe/entity:go_default_library",
         "//stovepipe/extension/buildrunner:go_default_library",
@@ -20,9 +20,9 @@ go_test(
     srcs = ["buildkite_test.go"],
     embed = [":go_default_library"],
     deps = [
+        "//platform/buildkite:go_default_library",
         "//platform/errs:go_default_library",
         "//platform/errs/http:go_default_library",
-        "//platform/extension/buildrunner/buildkite:go_default_library",
         "//platform/http:go_default_library",
         "//stovepipe/entity:go_default_library",
         "//stovepipe/extension/buildrunner:go_default_library",
