+++ stovepipe/extension/buildrunner/buildkite/buildkite_test.go
@@ -27,9 +27,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	platformbuildkite "github.com/uber/submitqueue/platform/buildkite"
 	"github.com/uber/submitqueue/platform/errs"
 	httperrs "github.com/uber/submitqueue/platform/errs/http"
-	platformbuildkite "github.com/uber/submitqueue/platform/extension/buildrunner/buildkite"
 	phttp "github.com/uber/submitqueue/platform/http"
 	"github.com/uber/submitqueue/stovepipe/entity"
 	"github.com/uber/submitqueue/stovepipe/extension/buildrunner"
