+++ submitqueue/extension/buildrunner/buildkite/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//platform/base/change:go_default_library",
-        "//platform/extension/buildrunner/buildkite:go_default_library",
+        "//platform/buildkite:go_default_library",
         "//submitqueue/core/changeset:go_default_library",
         "//submitqueue/entity:go_default_library",
         "//submitqueue/extension/buildrunner:go_default_library",
@@ -21,7 +21,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//platform/base/change:go_default_library",
-        "//platform/extension/buildrunner/buildkite:go_default_library",
+        "//platform/buildkite:go_default_library",
         "//platform/http:go_default_library",
         "//submitqueue/core/changeset:go_default_library",
         "//submitqueue/core/changeset/fake:go_default_library",
