+++ submitqueue/extension/buildrunner/buildkite/buildkite.go
@@ -37,7 +37,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/uber/submitqueue/platform/base/change"
-	platformbuildkite "github.com/uber/submitqueue/platform/extension/buildrunner/buildkite"
+	platformbuildkite "github.com/uber/submitqueue/platform/buildkite"
 	"github.com/uber/submitqueue/submitqueue/core/changeset"
 	"github.com/uber/submitqueue/submitqueue/entity"
 	"github.com/uber/submitqueue/submitqueue/extension/buildrunner"
